feat(dashboard): phase 4 — flip /network detail to v2 chrome
Per-network detail page (/network?id=…) moves from (dashboard) to
(v2-dashboard). Body strips PageContainer + Breadcrumbs + the legacy
Card.List info card; replaces them with the v2 page chrome and an
OzCard wrapping a divided list of network metadata (High availability
state + active policy count).

ResourcesSection + NetworkRoutingPeersSection stay unchanged — those
are self-contained sub-modules with their own state surfaces; a
deeper paint of each is tracked separately.

Edit button moves from the legacy pencil-on-link affordance to the
v2 7×7 icon button shape used everywhere else (Service Users, etc.).
Breadcrumb gains "Workspace > Networks > Network".
